projects/daily-podcast-bot/requirements.txt
daily-python>=0.11.0
openai>=1.30.0
websockets>=12.0
numpy>=1.26.0
sounddevice>=0.4.6
scipy>=1.12.0
python-dotenv>=1.0.0
